requirements.txt
Django==5.1.3
djangorestframework==3.15.2
djangorestframework_simplejwt==5.3.1
rest_framework_simplejwt==0.0.2
django-cors-headers==4.6.0
psycopg2==2.9.10
django-heroku==0.3.1
gunicorn==23.0.0